Liquidity Bot/requirements.txt
web3==6.11.0
ccxt==4.3.36
python-dotenv==1.0.1
gspread==5.12.4
oauth2client==4.1.3